LPD-85688 Moves classes to allow us reuse logic in other modules
