+++ modules/dxp/apps/osb/osb-faro/osb-faro-engine-client/src/main/java/com/liferay/osb/faro/engine/client/mixin/BaseMixin.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.osb.faro.web.internal.model.display.contacts.mixin;
+package com.liferay.osb.faro.engine.client.mixin;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
